refactor: moving more general components into a general area
